ai_agent/requirements.txt
fastapi
uvicorn[standard]
pydantic>=2.0.0,<3.0.0
requests
openai
langsmith
python-dotenv
tavily-python